[WIP]: imex and restructuring files.
