+++ dream/compressible/time/time.py
@@ -1,0 +1,15 @@
+from dream.time import TimeSchemes 
+
+# Proxy class, used as an intermediate interface.
+class CompressibleTimeSchemes(TimeSchemes, is_interface=True):
+    ...
+
+
+
+
+
+
+
+
+
+
